under way commit
complete task
- 각 Page에서 따로 만들어진 NavBar 중복 로직 및 Route 중복 방지를 위해 하나로 통일 (#issue8)
- Multi Page application로 구현된 로직 Single Page application로 구현하기(Router와 연동) (#issue7)
- JS Extensions => TS Extensions으로 변경 및 Type Error 해결하기 (#issue7)
- 정적 HTML file => handlebars file로 변경하기 (#isuue7)
