+++ public/cuisine.html
@@ -1,48 +1,43 @@
 <!DOCTYPE html>
 <html lang="ko">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<link rel="stylesheet" href="css/reset.css" />
+		<link rel="stylesheet" href="css/normalize.css" />
+		<link rel="stylesheet" href="css/cuisine.css" />
+		<link
+			rel="stylesheet"
+			href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css"
+			integrity="sha512-+4zCK9k+qNFUR5X+cKL9EIR+ZOhtIloNl9GIKS57V1MyNsYpYcUrUeQc9vNfzsWfV28IaLL3i96P9sdNyeRssA=="
+			crossorigin="anonymous"
+		/>
+		<title>cuisine</title>
+	</head>
 
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <link rel="stylesheet" href="css/reset.css" />
-  <link rel="stylesheet" href="css/normalize.css" />
-  <link rel="stylesheet" href="css/cuisine.css" />
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css"
-    integrity="sha512-+4zCK9k+qNFUR5X+cKL9EIR+ZOhtIloNl9GIKS57V1MyNsYpYcUrUeQc9vNfzsWfV28IaLL3i96P9sdNyeRssA=="
-    crossorigin="anonymous" />
-  <title>cuisine</title>
-</head>
-
-<body>
-  <div class="container">
-    <header>
-      요리보기
-    </header>
-    <button class="back-btn">뒤로</button>
-    <main class="main">
-      <div class="preview">
-
-        <ul class="preview-list"></ul>
-      </div>
-      <div class="container-wrap">
-      </div>
-      <div class="popup">
-        <p>
-          즐겨찾기에 </br>추가되었습니다.🍏
-        </p>
-      </div>
-      <div class="overlay"></div>
-    </main>
-    <nav>
-      <ul class="nav-list">
-        <li><a href="ingredient.html">요리재료 선택</a></li>
-        <li><a href="memo.html">장보기 메모</a></li>
-        <li><a href="recommend.html">오늘의 요리</a></li>
-        <li><a href="myinfo.html">회원 정보</a></li>
-      </ul>
-    </nav>
-  </div>
-  <script src="./js/cuisine.js"></script>
-</body>
-
+	<body>
+		<div class="container">
+			<header>요리보기</header>
+			<button class="back-btn">뒤로</button>
+			<main class="main">
+				<div class="preview">
+					<ul class="preview-list"></ul>
+				</div>
+				<div class="container-wrap"></div>
+				<div class="popup">
+					<p>즐겨찾기에 <br />추가되었습니다.🍏</p>
+				</div>
+				<div class="overlay"></div>
+			</main>
+			<nav>
+				<ul class="nav-list">
+					<li><a href="ingredient.html">요리재료 선택</a></li>
+					<li><a href="memo.html">장보기 메모</a></li>
+					<li><a href="recommend.html">오늘의 요리</a></li>
+					<li><a href="myinfo.html">회원 정보</a></li>
+				</ul>
+			</nav>
+		</div>
+		<script src="./js/cuisine.js"></script>
+	</body>
 </html>
