+++ public/memo.html
@@ -1,48 +1,48 @@
 <!DOCTYPE html>
 <html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./css/reset.css" />
-    <link rel="stylesheet" href="./css/normalize.css" />
-    <link rel="stylesheet" href="./css/memo.css" />
-    <script defer src="./js/memo.js"></script>
-    <title>memo</title>
-  </head>
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<link rel="stylesheet" href="./css/reset.css" />
+		<link rel="stylesheet" href="./css/normalize.css" />
+		<link rel="stylesheet" href="./css/memo.css" />
+		<script defer src="./js/memo.js"></script>
+		<title>memo</title>
+	</head>
 
-  <body>
-    <div class="container">
-      <header class="header-memo">
-        <h2>장보기메모</h2>
-      </header>
-      <main>
-        <section>
-          <h3>장보기 메모추가</h3>
-          <div class="cart-memo">
-            <input class="cart-memo-input" type="text" placeholder="내용을 입력해주세요" />
-            <ul class="cart-memo-list clearfix">
-              <!-- <li>
+	<body>
+		<div class="container">
+			<header class="header-memo">
+				<h2>장보기메모</h2>
+			</header>
+			<main>
+				<section>
+					<h3>장보기 메모추가</h3>
+					<div class="cart-memo">
+						<input class="cart-memo-input" type="text" placeholder="내용을 입력해주세요" />
+						<ul class="cart-memo-list clearfix">
+							<!-- <li>
               <input id="cart-item1" class="checkbox" type ="checkbox" name="">
               <label for ="cart-item1">소고기</label>
               <i class="remove">x</i>
             </li> -->
-            </ul>
-            <div class="cart-all-btns">
-              <input id="cart-all-check" class="checkbox" type="checkbox" />
-              <label for="cart-all-check">전체체크</label>
-              <button class="cart-all-remove"><i>x </i>전체삭제</button>
-            </div>
-          </div>
-        </section>
-      </main>
-      <nav>
-        <ul class="nav-list">
-          <li><a href="ingredient.html">요리재료 선택</a></li>
-          <li><a href="memo.html">장보기 메모</a></li>
-          <li><a href="recommend.html">오늘의 요리</a></li>
-          <li><a href="myinfo.html">회원 정보</a></li>
-        </ul>
-      </nav>
-    </div>
-  </body>
+						</ul>
+						<div class="cart-all-btns">
+							<input id="cart-all-check" class="checkbox" type="checkbox" />
+							<label for="cart-all-check">전체체크</label>
+							<button class="cart-all-remove"><i>x </i>전체삭제</button>
+						</div>
+					</div>
+				</section>
+			</main>
+			<nav>
+				<ul class="nav-list">
+					<li><a href="ingredient.html">요리재료 선택</a></li>
+					<li><a href="memo.html">장보기 메모</a></li>
+					<li><a href="recommend.html">오늘의 요리</a></li>
+					<li><a href="myinfo.html">회원 정보</a></li>
+				</ul>
+			</nav>
+		</div>
+	</body>
 </html>
