+++ src/css/login.css
@@ -23,7 +23,7 @@ body {
 .container {
 	height: 100vh;
 	background-color: var(--main-color);
-	overflow: hidden;
+	overflow: auto;
 }
 
 .container .login-inner {
