+++ src/css/navigationBar.css
@@ -1,0 +1,59 @@
+nav {
+	width: 100%;
+	height: 60px;
+	position: fixed;
+	left: 0;
+	bottom: 0;
+	z-index: 20;
+}
+.nav-list {
+	width: 100%;
+	height: 100%;
+	border-top: 2px solid #fff;
+	font-size: 0;
+}
+.nav-list > li {
+	color: #fff;
+	text-align: center;
+	display: inline-block;
+	box-sizing: border-box;
+	line-height: 60px;
+	background-color: var(--main-color);
+	width: calc(100% / 4);
+	font-size: 16px;
+	position: relative;
+}
+.nav-list a {
+	display: block;
+}
+.nav-list > li:hover {
+	color: #000;
+}
+.nav-list > li:not(:last-child)::after {
+	position: absolute;
+	top: 50%;
+	transform: translateY(-50%);
+	right: 0;
+	z-index: 100;
+	line-height: 60px;
+	content: '';
+	display: block;
+	width: 2px;
+	height: 80%;
+	background-color: #fff;
+}
+
+/* @media (min-width: 319px) and (max-width: 400px) { */
+.nav-list > li {
+	font-size: 14px;
+	line-height: 60px;
+	text-align: center;
+}
+.nav-list > li > a {
+	width: 60px;
+	display: inline-block;
+	vertical-align: middle;
+	line-height: 1.2;
+	word-spacing: 1.2rem;
+}
+/* } */
