+++ src/js/view/index.ts
@@ -1,6 +1,9 @@
 import 'CSS/reset.css';
 import 'CSS/normalize.css';
+import 'CSS/navigationBar.css';
 
 import { initialRoutes } from 'Router/router';
 
 initialRoutes();
+
+// window.onpopstate = pop
