+++ src/js/view/navigationBar.ts
@@ -1,0 +1,1 @@
+import 'CSS/navigationBar.css';
