+++ src/pages/cuisine.hbs
@@ -1,0 +1,23 @@
+<div class='container'>
+	<header>요리보기</header>
+	<button class='back-btn'>뒤로</button>
+	<main class='main'>
+		<div class='preview'>
+			<ul class='preview-list'></ul>
+		</div>
+		<div class='container-wrap'></div>
+		<div class='popup'>
+			<p>즐겨찾기에 <br />추가되었습니다.🍏</p>
+		</div>
+		<div class='overlay'></div>
+
+	</main>
+	{{! <nav>
+		<ul class='nav-list'>
+			<li><a href='ingredient.html'>요리재료 선택</a></li>
+			<li><a href='memo.html'>장보기 메모</a></li>
+			<li><a href='recommend.html'>오늘의 요리</a></li>
+			<li><a href='myinfo.html'>회원 정보</a></li>
+		</ul>
+	</nav> }}
+</div>
