+++ src/pages/ingredient.hbs
@@ -109,7 +109,7 @@
 
 		<div class='container-wrap'>
 			<div class='container-btn-wrap'>
-				<div class='cookEnter'>요리보기</div>
+				<div class='cookEnter' route='/cuisine'>요리보기</div>
 				<div class='refresh'><i class='fas fa-redo'></i></div>
 			</div>
 			<div class='ingre-item-container-wrap'>
@@ -119,12 +119,12 @@
 			</div>
 		</div>
 	</main>
-	<nav>
+	{{! <nav>
 		<ul class='nav-list'>
 			<li><a href='ingredient.html'>요리재료 선택</a></li>
 			<li><a href='memo.html'>장보기 메모</a></li>
 			<li><a href='recommend.html'>오늘의 요리</a></li>
 			<li><a href='myinfo.html'>회원 정보</a></li>
 		</ul>
-	</nav>
+	</nav> }}
 </div>
