+++ src/pages/navigationBar.hbs
@@ -1,0 +1,8 @@
+<nav>
+	<ul class='nav-list'>
+		<li><a href='ingredient.html'>요리재료 선택</a></li>
+		<li><a href='memo.html'>장보기 메모</a></li>
+		<li><a href='recommend.html'>오늘의 요리</a></li>
+		<li><a href='myinfo.html'>회원 정보</a></li>
+	</ul>
+</nav>
